### 2016-09-06 #56 @whedon
[REVIEW]: hddtools: Hydrological Data Discovery Tools
**Submitting author:** @cvitolo (<a href="http://orcid.org/0000-0002-4252-1176">Claudia Vitolo</a>)
**Repository:** <a href="https://github.com/ropensci/hddtools/" target ="_blank">https://github.com/ropensci/hddtools/</a>
**Version:** v0.3.0
**Archive:** <a href="https://dx.doi.org/10.5281/zenodo.247842" target="_blank">https://dx.doi.org/10.5281/zenodo.247842</a>

**Reviewers and authors**: Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue
in the <a href="https://github.com/cvitolo/hddtools/" target="_blank">target repository</a> and link to those issues (especially acceptance-blockers)
in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice
versa will create corresponding breadcrumb trails in the link target.)
