### 2016-09-07 #56 @cvitolo
@arfon I have decided to submit to ropensci as well. @noamross suggests to pause this review and wait for the one from ropensci. Would that be possible?
