### 2016-09-07 #56 @labarba
rOpenSci is not a journal. I'm not sure there is a conflict ...
