### 2016-09-07 #56 @noamross
@labarba There's not a conflict, but we're piloting a system where authors go through our review process, and on acceptance we submit their package to JOSS, and editors have the option of accepting based on rOpenSci's reviews rather than re-reviewing.
