### 2016-09-07 #56 @arfon
> @arfon I have decided to submit to ropensci as well. @noamross suggests to pause this review and wait for the one from ropensci. Would that be possible?

Yep it's OK to pause.

> rOpenSci is not a journal. I'm not sure there is a conflict ...

@labarba - this is something @noamross, @karthik and myself have been piloting. This review process is heavily based on the rOpenSci review process so JOSS reviews can go much faster if a package has already been through the rOpenSci review process.
