### 2017-01-16 #56 @noamross
This package has been reviewed and accepted by rOpenSci.  See the review thread here: https://github.com/ropensci/onboarding/issues/73 .

Note that, with repository transfer since the initial submission, the new repository URL is https://github.com/ropensci/hddtools . The current version is v0.5.0 and Zenodo DOI is http://doi.org/10.5281/zenodo.247842
