### 2017-01-20 #56 @arfon
> This package has been reviewed and accepted by rOpenSci. See the review thread here: ropensci/onboarding#73 .

👍 excellent, thanks @noamross. I'll proceed with accepting this submission now.
