### 2017-01-20 #56 @arfon
@cvitolo your paper is now accepted into JOSS. Your paper DOI is http://dx.doi.org/10.21105/joss.00056 ⚡️ 🚀 💥
