### 2017-04-04 #56 @surajitdb
Dear Claudia,
                   In Clark et al (2008) it was mentioned that 79 plausible model structures were constructed, however in your latest and previous tutorials using R states between 1 and 1248 (model identification number). Is there any upgradation of FUSE already?

I have been able run your latest tutorial successfully using R 3.2.5, but not working with R 3.3.3 due to some packages not available in the latest version. For now it is fine with R 3.2.5 version.

Could you suggest me some more mids apart from mids = 60(TOPMODEL), 230 (ARNOX/VIC), 343(PRMS) and 426(SACRAMENTO). along with their names, say 9 in all. I need 9 plausible model structures for my analysis.

How would one change the coding in R when we have to handle 9 model structures? I have copied snippet from your code.

bestModel <- function(runNumber){

  if (runNumber < (numberOfRuns + 1)) myBestModel <- "TOPMODEL"

  if (runNumber > (numberOfRuns + 1) &
      runNumber < (2*numberOfRuns + 1)) myBestModel <- "ARNOXVIC"

  if (runNumber > (2*numberOfRuns + 1) &
      runNumber < (3*numberOfRuns + 1)) myBestModel <- "PRMS"

  if (runNumber > (3*numberOfRuns + 1) &
      runNumber < (4*numberOfRuns + 1)) myBestModel <- "SACRAMENTO"
if (runNumber > (4*numberOfRuns +1) & runNumber < (5*numberOfRuns + 1)) myBestModel <- "Another model"
# Does the above code correct if I have 5 model structures for my analysis?

  return(myBestModel)

}

Which model structure is being calibrated using hydromad?

Looking forward to your response in anticipation.

Thanking you.

Regards,

Surajit
