### 2017-04-04 #56 @surajitdb
This was tested according to https://cvitolo.github.io/fuse/articles/fuse_vignette.html.
