### 2017-04-04 #56 @surajitdb
Sorry Claudia,
                      The snippet of the code is
How would one change the coding in R when we have to handle 9 model structures? I have copied snippet from your code.

bestModel <- function(runNumber){

if (runNumber < (numberOfRuns + 1)) myBestModel <- "TOPMODEL"

if (runNumber > (numberOfRuns + 1) &
runNumber < (2*numberOfRuns + 1)) myBestModel <- "ARNOXVIC"

if (runNumber > (2*numberOfRuns + 1) &
runNumber < (3*numberOfRuns + 1)) myBestModel <- "PRMS"

if (runNumber > (3*numberOfRuns + 1) &
runNumber < (4*numberOfRuns + 1)) myBestModel <- "SACRAMENTO"
if (runNumber > (4*numberOfRuns +1) & runNumber < (5*numberOfRuns + 1)) myBestModel <- "Another model"

return(myBestModel)

}
